requirements.txt
asgiref==3.7.2
attrs==23.2.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
s3cmd==2.4.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
whitenoise==6.6.0
